Revert "Added blocked, spam functionality (#1701)"
This reverts commit bd27aa52ae8ec71783f090ee00f6d9d982574616.
