+++ packages/chat-sdk/src/components/ChatContext.tsx
@@ -14,11 +14,6 @@ type ChatContext = {
   requested: boolean;
   remoteUserId: string;
   type: SubscriptionType;
-  blocked?: boolean;
-  spam?: boolean;
-  setRequested?: any;
-  setSpam?: any;
-  setBlocked?: any;
 };
 
 export const _ChatContext = React.createContext<ChatContext | null>(null);
@@ -36,11 +31,6 @@ export function ChatProvider(props: {
   requested: boolean;
   remoteUserId: string;
   type: SubscriptionType;
-  blocked?: boolean;
-  spam?: boolean;
-  setRequested?: any;
-  setSpam?: any;
-  setBlocked?: any;
 }) {
   return (
     <_ChatContext.Provider
@@ -56,11 +46,6 @@ export function ChatProvider(props: {
         requested: props.requested,
         remoteUserId: props.remoteUserId,
         type: props.type,
-        spam: props.spam,
-        blocked: props.blocked,
-        setRequested: props.setRequested,
-        setSpam: props.setSpam,
-        setBlocked: props.setBlocked,
       }}
     >
       {props.children}
