+++ packages/chat-sdk/src/components/ChatRoom.tsx
@@ -14,11 +14,6 @@ interface ChatRoomProps {
   areFriends?: boolean;
   requested?: boolean;
   remoteUserId?: string;
-  blocked?: boolean;
-  spam?: boolean;
-  setRequested?: any;
-  setSpam?: any;
-  setBlocked?: any;
 }
 
 export const ChatRoom = ({
@@ -30,11 +25,6 @@ export const ChatRoom = ({
   areFriends = true,
   requested = false,
   remoteUserId,
-  blocked,
-  spam,
-  setRequested,
-  setSpam,
-  setBlocked,
 }: ChatRoomProps) => {
   const [chatManager, setChatManager] = useState<ChatManager | null>(null);
   const messageContainerRef = useRef(null);
@@ -98,11 +88,6 @@ export const ChatRoom = ({
       requested={requested}
       remoteUserId={remoteUserId || ""}
       type={type}
-      spam={spam}
-      blocked={blocked}
-      setRequested={setRequested}
-      setSpam={setSpam}
-      setBlocked={setBlocked}
     >
       <FullScreenChat />
     </ChatProvider>
